fix: remove SearchRepositoryImplTest due to wrong location
